puppeteer

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -1,6 +1,7 @@
 # excecute kill me now on loop bash script
 exec { 'killmenow':
-  command     => 'pkill killmenow',
-  path        => '/usr/bin:/usr/sbin:/bin',
-  refreshonly => true,
+  command => 'pkill killmenow',
+  path    => '/usr/bin',
+  onlyif  => 'pgrep killmenow',
 }
+
